Have claim call resurrect on ENTER key press
