+++ src/features/embalm/stepContent/hooks/useEnterToNextStep.ts
@@ -1,4 +1,4 @@
-import { useEffect } from 'react';
+import { useEnterKeyCallback } from 'hooks/useEnterKeyCallback';
 import { Step } from 'store/embalm/reducer';
 import { useSarcophagusParameters } from './useSarcophagusParameters';
 import { useStepContent } from './useStepContent';
@@ -11,7 +11,7 @@ export function useEnterToNextStep() {
   const { goToNext, currentStep } = useStepContent();
   const { sarcophagusParameters } = useSarcophagusParameters();
 
-  useEffect(() => {
+  useEnterKeyCallback(() => {
     if (currentStep === Step.CreateSarcophagus) return;
 
     const currentStepParams = sarcophagusParameters.find(s => s.step === currentStep);
@@ -19,20 +19,12 @@ export function useEnterToNextStep() {
     const canGoNext =
       currentStep === Step.NameSarcophagus
         ? !currentStepParams?.error &&
-          !sarcophagusParameters.find(s => s.name === 'RESURRECTION')?.error
+        !sarcophagusParameters.find(s => s.name === 'RESURRECTION')?.error
         : !currentStepParams?.error;
 
-    const keyDownHandler = (event: any) => {
-      if (event.key === 'Enter' && canGoNext) {
-        event.preventDefault();
-        goToNext();
-      }
-    };
 
-    document.addEventListener('keydown', keyDownHandler);
-
-    return () => {
-      document.removeEventListener('keydown', keyDownHandler);
-    };
-  }, [goToNext, currentStep, sarcophagusParameters]);
+    if (canGoNext) {
+      goToNext();
+    }
+  });
 }
